requirements.txt
django==3.1
djangorestframework
psycopg2-binary==2.8.3
celery==5.0.1
Pillow
markdown
django-filter